Update the guide and code
