+++ content/guides/github-action-integration/code/workflow.yml
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-//#workflow-example
 name: Run Gatling Enterprise Simulation
 
 on:
